[ttx_diff] Add test_helpers module to layout-normalizer
